Added resources update for conversion and upgrade

+++ Assets/Scripts/Networking/Headquarters/PlayerData.cs
@@ -9,7 +9,7 @@ namespace Networking.Headquarters
     {
         public string Id { get; set; }
         public byte Level { get; set; }
-        public ushort Experience { get; set; }
+        public uint Experience { get; set; }
         public uint Energy { get; set; }
         public Resource[] Resources { get; set; }
         public Robot[] Robots { get; set; }
@@ -19,7 +19,7 @@ namespace Networking.Headquarters
 
         public PlayerData() { }
 
-        public PlayerData(string id, byte level, ushort experience, uint energy, Resource[] resources, Robot[] robots,
+        public PlayerData(string id, byte level, uint experience, uint energy, Resource[] resources, Robot[] robots,
             BuildTask[] conversionQueue, BuildTask[] upgradeQueue, BuildTask[] buildQueue)
         {
             Id = id;
@@ -41,7 +41,7 @@ namespace Networking.Headquarters
         {
             Id = e.Reader.ReadString();
             Level = e.Reader.ReadByte();
-            Experience = e.Reader.ReadUInt16();
+            Experience = e.Reader.ReadUInt32();
             Energy = e.Reader.ReadUInt32();
             Resources = e.Reader.ReadSerializables<Resource>();
             Robots = e.Reader.ReadSerializables<Robot>();
